Refactor: 게시글 비교 로직을 `crawling`으로 책임 이전
