+++ src/main/java/com/team7/retriever/domain/crawling/controller/dto/request/UpdateCheckRequest.java
@@ -1,0 +1,8 @@
+package com.team7.retriever.domain.crawling.controller.dto.request;
+
+public record UpdateCheckRequest(
+	String link,
+	String title,
+	String source
+) {
+}
