+++ src/main/java/com/team7/retriever/domain/crawling/controller/dto/request/UpdateRequest.java
@@ -1,4 +1,4 @@
-package com.team7.retriever.domain.post.controller.dto.request;
+package com.team7.retriever.domain.crawling.controller.dto.request;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @Builder
-public class PostUpdateRequest {
+public class UpdateRequest {
 	private String html;
 	private String link;
 	private String title;
