--- src/main/java/com/team7/retriever/domain/post/controller/dto/request/PostUpdateCheckRequest.java
@@ -1,8 +1,0 @@
-package com.team7.retriever.domain.post.controller.dto.request;
-
-public record PostUpdateCheckRequest(
-	String link,
-	String title,
-	String source
-) {
-}
